Apply list info in redux
