+++ package.json
@@ -9,7 +9,8 @@
     "react-redux": "^7.0.3",
     "react-scripts": "3.0.1",
     "reactstrap": "^8.0.0",
-    "redux": "^4.0.1"
+    "redux": "^4.0.1",
+    "redux-logger": "^3.0.6"
   },
   "scripts": {
     "start": "react-scripts start",
