+++ src/redux/initActions/types.js
@@ -1,0 +1,4 @@
+export const CLICK_SUB = 'CLICK_SUB';
+export const CLICK_ADD = 'CLICK_ADD';
+export const CLICK_CHECKOUT = 'CLICK_CHECKOUT';
+export const CLICK_RESET = 'CLICK_RESET';
