+++ src/redux/initReducers/titles.js
@@ -1,0 +1,11 @@
+const titles = [{id: 1, title: 'Cold cuts', price: 5},
+    {id: 2, title: 'Pepperoni', price: 3.5},
+    {id: 3, title: 'Feta', price: 2.5},
+    {id: 4, title: 'Mozzarella', price: 1.5},
+    {id: 5, title: 'Swiss cheese', price: 3},
+    {id: 6, title: 'Spices', price: 0.5},
+    {id: 7, title: 'Vegetables', price: 1.25}];
+
+export default function (state = titles, action) {
+    return state;
+}
